# ARROW-13238: [C++][Compute][Dataset] Use an ExecPlan for dataset scans