… cpp/src/arrow/compute/exec/plan_test.cc
… cpp/src/arrow/compute/exec_test.cc
… cpp/src/arrow/dataset/file_parquet_test.cc
… cpp/src/arrow/dataset/scanner_test.cc
… cpp/src/arrow/testing/future_util.h
… cpp/src/arrow/testing/matchers.h
… cpp/src/arrow/util/async_generator.h
… cpp/src/arrow/util/future.cc
… cpp/src/arrow/util/future.h
… cpp/src/arrow/util/future_test.cc
… cpp/src/arrow/util/thread_pool.h
… dev/archery/archery/lang/cpp.py