Replaces the body of AsyncScanner::ScanBatchesAsync with usage of an ExecPlan