@lidavidm 2021-07-06
I'll test this on EC2 when I get a chance and compare to 4.0/current master.