@lidavidm 2021-07-06
This got stuck while I was testing it (I'll try to get a backtrace shortly). But performance indeed seems to be on par with before.

<details>

4.0.1:
```
        "iterations": 10,
        "max": "10.501172",
        "mean": "9.783475",
        "median": "9.783779",
        "min": "9.203580",
        "q1": "9.451842",
        "q3": "10.074056",
```

5.0.0 master
```
        "iterations": 10,
        "max": "12.266335",
        "mean": "10.373638",
        "median": "10.223064",
        "min": "9.785621",
        "q1": "10.084575",
        "q3": "10.322508",
```
Count rows, no filter, 10 iterations:
```
Min: 0.39461345200015785 s
Median: 0.420534689500073 s
Mean: 0.4404416943999422 s
Max: 0.5733599779996439 s
```

This PR
```
        "iterations": 10,
        "max": "11.053488",
        "mean": "10.406635",
        "median": "10.273359",
        "min": "9.980684",
        "q1": "10.145428",
        "q3": "10.701346",
```
Count rows, no filter, 10 iterations:
```
Min: 0.3555871419994219 s
Median: 0.38891541300017707 s
Mean: 0.4031996126002014 s
Max: 0.5227700120003647 s
```
(It got stuck when a filter was applied)

<details>