@lidavidm 2021-07-06
Now that I'm running it again it just seems that count_rows with a filter got very slow.

```
Before:
Min: 24.285114823000185 s
Median: 25.309564675000274 s
Mean: 25.45627587750014 s
Max: 26.564001424999333 s

After:
Min: 113.42360810700029 s
Median: 117.11463808250028 s
Mean: 118.39458996600005 s
Max: 123.9104565839998 s
```

<details>
<summary>Test script</summary>

```python
import statistics
import time
import pyarrow 
import pyarrow.dataset
import pyarrow.fs

fs = pyarrow.fs.S3FileSystem(region="us-east-2")

print("PyArrow:", pyarrow.__version__)

times = []
for _ in range(10):
    start = time.monotonic()
    ds = pyarrow.dataset.dataset([
        "ursa-labs-taxi-data/2009/01/data.parquet",
        "ursa-labs-taxi-data/2009/02/data.parquet",
        "ursa-labs-taxi-data/2009/03/data.parquet",
        "ursa-labs-taxi-data/2009/04/data.parquet",
        ],
        format="parquet",
        partitioning=["year", "month"],
        filesystem=fs,
    )
    print("Rows:", ds.scanner(use_async=True, use_threads=True).count_rows())
    times.append(time.monotonic() - start)

print("No filter")
print("Min:", min(times), "s")
print("Median:", statistics.median(times), "s")
print("Mean:", statistics.mean(times), "s")
print("Max:", max(times), "s")

print("With filter")
times = []
expr = pyarrow.dataset.field("passenger_count") > 1
for _ in range(10):
    start = time.monotonic()
    ds = pyarrow.dataset.dataset([
        "ursa-labs-taxi-data/2009/01/data.parquet",
        "ursa-labs-taxi-data/2009/02/data.parquet",
        "ursa-labs-taxi-data/2009/03/data.parquet",
        "ursa-labs-taxi-data/2009/04/data.parquet",
        ],
        format="parquet",
        partitioning=["year", "month"],
        filesystem=fs,
    )
    print("Rows:", ds.scanner(use_async=True, use_threads=True, filter=expr).count_rows())
    times.append(time.monotonic() - start)
print("Min:", min(times), "s")
print("Median:", statistics.median(times), "s")
print("Mean:", statistics.mean(times), "s")
print("Max:", max(times), "s")
```

</details>