@bkietz 2021-07-07
Well that's disappointing. I was actually expecting it to get faster