@lidavidm 2021-07-07
If we correct the CountRows issue above, CountRows indeed gets much faster!

```
Before:

Min: 24.285114823000185 s
Median: 25.309564675000274 s
Mean: 25.45627587750014 s
Max: 26.564001424999333 s

After:

Min: 6.617772240000249 s
Median: 6.906719556000098 s
Mean: 6.926710154500052 s
Max: 7.311853507000251 s
```